Fix bug for connecting to MariaDB's database.
Since Laravel 5.4, the default characater set is changed to utf8mb4
this has no problem when dealing with recents versions of MySQL.

However it has problems when dealing with MariaDB, so a fix to this
was found in this site.
https://laravel-news.com/laravel-5-4-key-too-long-error

By adding a String that sets the String lenght the problem was fixed.