Add Vue.js front end first implementation to the admin view.
This was a big commit for me. First of all, the create view is no longer use and
it now part of the admin view, toggled by the create service button.
Furthermore, the example script for the Place's Autocomplete API with Vue.js wasn't
working. So I rewrite the behaviour into Vue.js, reference the autocomplete input
and fetch the data from the API to the Vue.js model binding.